third_party/blink/web_tests/external/wpt/shadow-dom/declarative/declarative-shadow-dom-basic-expected.txt

This is a testharness.js-based test.
[FAIL] Shadowrootmode reflection
  assert_equals: Invalid values map to empty string expected (string) "" but got (object) null
[FAIL] Shadowrootmode reflection, setter
  assert_equals: expected (string) "" but got (object) null
Harness: the test ran to completion.
